add auth helpers
