+++ package.json
@@ -22,15 +22,18 @@
     },
     "dependencies": {
         "@hono/zod-validator": "0.4.3",
-        "drizzle-kit": "0.30.6",
+        "@oslojs/crypto": "1.0.1",
+        "@oslojs/encoding": "1.1.0",
         "drizzle-orm": "0.41.0",
-        "hono": "4.7.5",
-        "mysql2": "3.14.0",
+        "drizzle-kit": "0.30.6",
         "winston": "3.17.0",
+        "mysql2": "3.14.0",
+        "hono": "4.7.5",
         "zod": "3.24.2"
     },
     "devDependencies": {
         "@types/bun": "1.2.8",
-        "prettier": "3.5.3"
+        "prettier": "3.5.3",
+        "drizzle-seed": "0.3.1"
     }
 }
